requirements.txt
psycopg2-binary
shapely
geopandas
sqlalchemy
